+++ src/calc-size-fast.h
@@ -1,0 +1,2 @@
+
+int calc_size_fast_outer(SEXP s);