+++ src/calc-size-only.h
@@ -1,2 +1,0 @@
-
-int calc_marshalled_size(SEXP robj);