+++ src/calc-size-robust.c
@@ -38,7 +38,7 @@ void count_bytes(R_outpstream_t stream, void *src, int length) {
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 // Serialize an R object, but ony count the bytes.  C function
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-int calc_marshalled_size(SEXP robj) {
+int calc_size_robust(SEXP robj) {
 
   // Initialise the count
   int count = 0;
@@ -69,6 +69,6 @@ int calc_marshalled_size(SEXP robj) {
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 // Serialize an R object, but ony count the bytes. R shim function
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-SEXP calc_marshalled_size_(SEXP robj) {
-  return ScalarInteger(calc_marshalled_size(robj));
+SEXP calc_size_robust_(SEXP robj) {
+  return ScalarInteger(calc_size_robust(robj));
 }