+++ src/init.c
@@ -4,15 +4,17 @@
 
 extern SEXP marshall_();
 extern SEXP unmarshall_();
-extern SEXP calc_marshalled_size_();
-extern SEXP marshall_minimize_malloc_();
+extern SEXP calc_size_robust_();
+extern SEXP calc_size_fast_();
+extern SEXP marshall_fast_();
 
 static const R_CallMethodDef CEntries[] = {
 
-  {"marshall_"  , (DL_FUNC) &marshall_  , 1},
-  {"unmarshall_", (DL_FUNC) &unmarshall_, 1},
-  {"calc_marshalled_size_", (DL_FUNC) &calc_marshalled_size_, 1},
-  {"marshall_minimize_malloc_", (DL_FUNC) &marshall_minimize_malloc_, 1},
+  {"marshall_"        , (DL_FUNC) &marshall_            , 1},
+  {"unmarshall_"      , (DL_FUNC) &unmarshall_          , 1},
+  {"calc_size_robust_", (DL_FUNC) &calc_size_robust_    , 1},
+  {"calc_size_fast_"  , (DL_FUNC) &calc_size_fast_      , 1},
+  {"marshall_fast_"   , (DL_FUNC) &marshall_fast_       , 2},
   {NULL , NULL, 0}
 };